backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-dotenv==1.0.1
python-multipart==0.0.9
langchain-openai==0.2.0
langchain==0.3.0
sentence-transformers==3.1.0
PyMuPDF==1.24.10
faiss-cpu==1.8.0
numpy==1.26.4
pydantic==2.9.0
